+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/ir.kt
@@ -770,44 +770,3 @@ internal fun IrFunction.stubExpressionBody(context: IrMetroContext) =
 
 internal fun IrBuilderWithScope.stubExpression(context: IrMetroContext) =
   irInvoke(callee = context.symbols.stdlibErrorFunction, args = listOf(irString("Never called")))
-
-internal fun IrMetroContext.trackFunctionCall(
-  callingDeclaration: IrDeclaration,
-  calleeFunction: IrFunction,
-) {
-  val ktFile = callingDeclaration.fileParentOrNull?.getKtFile()
-  if ((ktFile != null && ktFile.doNotAnalyze == null) || ktFile == null) {
-    // Not every declaration has a file parent, for example IR-generated accessors
-    val filePath =
-      ktFile?.virtualFile?.path ?: callingDeclaration.fileParentOrNull?.fileEntry?.name ?: return
-    val declaration =
-      (calleeFunction as? IrSimpleFunction)?.correspondingPropertySymbol?.owner ?: calleeFunction
-    trackLookup(
-      container = calleeFunction.parentAsClass.kotlinFqName,
-      functionName = declaration.name.asString(),
-      location =
-        object : LocationInfo {
-          override val filePath = filePath
-          override val position: Position = Position.NO_POSITION
-        },
-    )
-  }
-}
-
-internal fun IrMetroContext.trackLookup(
-  container: FqName,
-  functionName: String,
-  location: LocationInfo,
-) {
-  lookupTracker?.let { tracker ->
-    synchronized(tracker) {
-      tracker.record(
-        filePath = location.filePath,
-        position = if (tracker.requiresPosition) location.position else Position.NO_POSITION,
-        scopeFqName = container.asString(),
-        scopeKind = ScopeKind.CLASSIFIER,
-        name = functionName,
-      )
-    }
-  }
-}
